Found and fixed device rotation values
It turns out that Input.Gyro.Attitude gives rotation values of device relative to the space its in.
Because of this i changed the script to use Input.Gyro.rotationRateUnbiased, which gives the change in rotation values from the device.
Because of this the calibration is no longer necessary as the user can tilt the device as he/she wants, since it returns acceleration now, not orientation.